+++ tests/profiling/collector/test_threading.py
@@ -2215,33 +2215,3 @@ def test_exclude_modules_multiple() -> None:
     with ThreadingLockCollector(capture_pct=100):
         lock = threading.Lock()
         assert not isinstance(lock, _ProfiledLock), "Lock from excluded module should be native"
-
-
-class TestExcludeModulesConfig:
-    """Unit tests for the exclude_modules config field type guarantees."""
-
-    def test_default_is_empty_frozenset(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """exclude_modules must default to frozenset(), not '' or None."""
-        monkeypatch.delenv("DD_PROFILING_LOCK_EXCLUDE_MODULES", raising=False)
-        from ddtrace.internal.settings.profiling import ProfilingConfigLock
-
-        cfg = ProfilingConfigLock()
-        assert cfg.exclude_modules == frozenset()
-        assert isinstance(cfg.exclude_modules, frozenset)
-
-    def test_parsed_value_is_frozenset(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """A non-empty env var must produce a frozenset[str], not a raw string."""
-        monkeypatch.setenv("DD_PROFILING_LOCK_EXCLUDE_MODULES", "uvicorn,asyncio,sqlalchemy.pool")
-        from ddtrace.internal.settings.profiling import ProfilingConfigLock
-
-        cfg = ProfilingConfigLock()
-        assert isinstance(cfg.exclude_modules, frozenset)
-        assert cfg.exclude_modules == frozenset({"uvicorn", "asyncio", "sqlalchemy.pool"})
-
-    def test_whitespace_stripped(self, monkeypatch: pytest.MonkeyPatch) -> None:
-        """Leading/trailing whitespace around module names must be stripped."""
-        monkeypatch.setenv("DD_PROFILING_LOCK_EXCLUDE_MODULES", " uvicorn , asyncio ")
-        from ddtrace.internal.settings.profiling import ProfilingConfigLock
-
-        cfg = ProfilingConfigLock()
-        assert cfg.exclude_modules == frozenset({"uvicorn", "asyncio"})
