+++ tests/profiling/test_profiling_config.py
@@ -48,3 +48,32 @@ class TestAdaptiveSamplingConfig:
         monkeypatch.setenv("_DD_PROFILING_STACK_ADAPTIVE_SAMPLING_MAX_INTERVAL_US", "2000000")
         with pytest.raises(ValueError):
             ProfilingConfig()
+
+
+class TestExcludeModulesConfig:
+    """Unit tests for the exclude_modules config field type guarantees."""
+
+    def test_default_is_empty_set(self, monkeypatch: pytest.MonkeyPatch) -> None:
+        """exclude_modules must default to frozenset(), not '' or None."""
+        monkeypatch.delenv("DD_PROFILING_LOCK_EXCLUDE_MODULES", raising=False)
+        from ddtrace.internal.settings.profiling import ProfilingConfigLock
+
+        cfg = ProfilingConfigLock()
+        assert isinstance(cfg.exclude_modules, set)
+
+    def test_parsed_value_is_frozenset(self, monkeypatch: pytest.MonkeyPatch) -> None:
+        """A non-empty env var must produce a frozenset[str], not a raw string."""
+        monkeypatch.setenv("DD_PROFILING_LOCK_EXCLUDE_MODULES", "uvicorn,asyncio,sqlalchemy.pool")
+        from ddtrace.internal.settings.profiling import ProfilingConfigLock
+
+        cfg = ProfilingConfigLock()
+        assert isinstance(cfg.exclude_modules, set)
+        assert cfg.exclude_modules == set({"uvicorn", "asyncio", "sqlalchemy.pool"})
+
+    def test_whitespace_stripped(self, monkeypatch: pytest.MonkeyPatch) -> None:
+        """Leading/trailing whitespace around module names must be stripped."""
+        monkeypatch.setenv("DD_PROFILING_LOCK_EXCLUDE_MODULES", " uvicorn , asyncio ")
+        from ddtrace.internal.settings.profiling import ProfilingConfigLock
+
+        cfg = ProfilingConfigLock()
+        assert cfg.exclude_modules == set({"uvicorn", "asyncio"})
